All four reset functions armed; flow cookie works; cause string REGRESSED
WHAT WORKS, verified live on tmm:all3 with all four addresses armed on both pods:

    records  flows  site
         15     15  http_mr_proxy.c:993
         15     15  http_mr_proxy.c:994
          5      2  flow_table.c:2618
          2      2  tcp.c:4689

THE FLOW COOKIE DOES EXACTLY WHAT IT WAS ADDED FOR. flow_table.c shows 5 records from
only 2 DISTINCT FLOWS, while http_mr_proxy.c shows 15 across 15. That is the
cardinality distinction --- one client hammering versus systemic --- and it is visible
at a glance instead of being unanswerable. "flow":"00003ef7f8866cb5" in the JSON.

ALL FOUR FUNCTIONS ARM. I was wrong earlier that varargs was "the real limit, not a
small fix". The signatures settle it:

    rst_why         (uf, file, lineno, err, reason, cause)
    rst_why_va      (uf, file, lineno, err, reason, cause, fmt, ...)
    rst_why_preserve(uf, file, lineno, err, cause)

rst_why_va's first SIX arguments are identical to rst_why's --- varargs begin at the
seventh, and the dispatcher reads only six --- so it shares the builder verbatim and
needed only its own slot. rst_why_preserve has no `reason`, so the cause is a4 not a5;
that is the single shape difference and why it gets slot 3, shared with
rst_why_preserve_va. Four addresses, three slots, 1,116 of 1,116 call sites.

THE REGRESSION, reported precisely rather than guessed at. "cause":"" on every record.
Hand-decoded from the ring bytes:

    cookie_lo/hi = 00003b17873633a9   correct, from a0
    lineno = 4689, err = 32, reason = 0, file_len = 5, file = "tcp.c"   all correct
    cause_len = 0, cause[36] = all zeros

grep -a finds no cause text anywhere in the segment, so the builder received NULL. a0,
a2, a3 and a4 all arrive correctly, so ONLY a5/r9 is wrong. It worked in the schema-2
build, so adding the cookie caused it --- prime suspect is the 7th parameter to
ls_ctx_rst_build with ls_uflow_cookie(a0) evaluated inside the argument list.

NOT FIXED, and not to be guessed at further: the next step is to carry raw a4/a5 in the
record for one build, which distinguishes "the trampoline's r9 is wrong" from "the
builder's parameter is wrong". I have been alternating between those two without
evidence, which is how the day's other mistakes happened.

Two void diagnostics worth recording so they are not repeated: `strings` is not
installed in the f5-tmm container (my own runbook says so) and returned 0 for text that
demonstrably decodes; and the first isolation test disarmed the preserve slots, which
ruled out the a4/a5 shift but not the builder.

+++ substrate/ls_flow_cookie.h
@@ -7,11 +7,16 @@
 #ifndef LS_FLOW_COOKIE_H
 #define LS_FLOW_COOKIE_H
 
-#include <stdint.h>
+/* NO #include, deliberately. This header is read from BOTH include worlds: from
+ * ls_tramp.c (STDINC, standard C) and from ls_flow_cookie.c (TMM's -nostdinc world,
+ * where <stdint.h> resolves to the system header and dies on
+ * bits/libc-header-start.h). `unsigned long long` is a builtin and needs no header,
+ * so it is the only portable choice across the boundary. It is >= 64 bits by the
+ * C standard, which is all the cookie requires. */
 
 /* TMM's own cookie for this flow, or 0 when there is no flow. Zero is a legitimate
  * answer, not a failure --- flow_table.c rejects flows before one exists, and
  * rst_why() itself reports those with a zero cookie. */
-uint64_t ls_uflow_cookie(void *uf);
+unsigned long long ls_uflow_cookie(void *uf);
 
 #endif /* LS_FLOW_COOKIE_H */
